Distinguish more clearly between the three states.
Showing the things you need to fix in green was surprising to me.

Errors and successfully determining that a project is stale both deserve
to be red, but use bold so that we can see the projects that need to be
updated even while there are lots of unsupported METADATA files.

Bug: N/A
Test: manual
Change-Id: Idad71a46c675b19d1fa204268c27816375f3e394