[Updater] Ignore commits brought in by a merge
As an example, if there is commit C (300 days old) in dev branch. Later dev is merged into upstream-master as commit M (2 days old).

Before this change, the updater will say "2 commits (300 days) behind" because both C and M is not in master.

After this change, it will say "1 commit (2 days) behind", which makes more sense.

(This change raises error when there's no merge base between upstream and master. But I think it is expected)

Test: updater.sh check googletest
Change-Id: I963b378e61677c1f6ea13af0d0f558846040b840