Mark CLs as Verified-1 if they have errors.
We currently upload CLs that can be broken.  This is potentially fine,
as it's a signal to the reviewer that an upgrade needs work.  But we
should mark them as failing in some way to give the reviewer that
hint.  This CL does that by setting Verified-1.

We currently only implement it for Rust crates.  Rather than modifying
the existing mechanisms to detect various errors (which seems
difficult e.g., in the case of patches) we do checks for errors after
the fact.  We currently check for .rej files (meaning failed patches)
and errors in Android.bp.  These checks are both heuristics and so
could fail (in either way).

Test: Created aosp/1665704, aosp/1667706, aosp/1667707
Change-Id: I7e148d4d0cc5c1f2434cfdb98a969e7969a3893a