Add tests for command line path resolution.
I'm about to rewrite this code to fix http://b/243685332, and wanted
to get tests written so I understood what I was about to change.

I haven't tested the glob behavior, because that's about to be removed
(it's actually a partial cause of the bug, and not needed).

The @cache was removed from external_path() because it made it
impossible to use a temp directory for that path in tests, because it
would be cached by an earlier test. It's not really an expensive path
anyway.

The code itself was moved from external_updater.py to fileutils.py
because Soong won't let us test anything in that file (because it's
python_binary, not python_library). It was also renamed for clarity.

Bug: http://b/243685332
Test: atest --host-unit-test-only
Change-Id: I63239a05448dbcc2b55c1d7274cfed2a0ed352b6