Remove mangling of command line paths.
Previously paths passed on the command line would be resolved as
either absolute (though not explicitly in the code, that was a quirk
of `a / b` just returning `b` if it was absolute), or relative to
//external. That meant that relatives paths people expected to work
like `external/foo`, `../foo`, `.`, etc. did not work, since those
were resolved relative to `//external` rather than the CWD. It doesn't
sound like anyone expected that behavior, so just remove it.

The other quirk that was caused by that is now removed is that,
because the path was relative to external, which the shell couldn't
know, external_updater had to do glob expansion rather than the shell,
leading to awkward commands like `update "rust/*"` instead of the more
obvious `update external/rust/*`.

There are still more callers to get_absolute_project_path() and
external_path() that probably need to be updated as well, but this
fixes the case that's most user-visible (tab completion resulting in
the wrong arguments).

Bug: http://b/243685332
Test: tests added/updated to reflect new behavior
Test: ./updater.sh check ../../external/rust/crates/*
Test: tools/external_updater/updater.sh check external/rust/crates/*
Test: ./updater.sh update ../external/fonttools # from //bionic
Test: /path/to/updater.sh . # in a different tree
Change-Id: Ia6370be268b232bac145cf8000b955b72485cafc