Run the build with the full path.
Previously this assumed that the CWD was the root of the tree
containing the project. That was wrong if the project was in a
different tree than external_updater (rare), but also requiring the
CWD to be the root of the tree makes tab completed paths from the CLI
not work unless the user happened to be in the root of the tree.

Bug: http://b/243685332
Test: ./updater.sh update toybox
Change-Id: I06826900937ea37bcc02e208d5d1024945e01c06