Add test coverage for git_utils.is_ancestor.
The GitRepo module is moved up to the top level of the tests package
so it can be used by tests outside endtoend (we technically could use
it without moving it, but we can't use the rest of the package, so
it'd be odd). I haven't added test_gitrepo.py to the Android.bp yet
because it's still written for pytest, which soong can't handle. I'll
follow up with another CL to rewrite that test to use the old test
runner.

Bug: None
Test: This is the test
Change-Id: I016daf4616efec10773dc23dec8f8cd01cb77707