Add a Makefile for scripting development tasks.
Neither of these goals actually pass currently because both are
currently broken...

The fact that we don't currently have a requirements.txt make the venv
management a bit more annoying than it should be. Managing those by
hand is annoying, so I'd like to check this in while we decide whether
to do it that way or use something that'll do it for us like Poetry
(but then you've got a different problem, although hopefully a smaller
problem).

Also currently absent is formatting, since there doesn't appear to be
a single code style here. That could be added easily if we decide on
what tools to use.

I haven't added code coverage for the test goal because it was acting
strangely and including repo's source in the coverage checks for some
reason.

Bug: None
Test: make check
Change-Id: If0910a9f70451748c22c7582ac9095c2e1324d38