Suppress broad-exception-caught lint.
Unclear whether this should be cleaned up or not. Figuring that out
would require a lot of digging to figure out which kinds of exceptions
this is trying to handle, if not all. If the goal is to catch all so
that `validate` doesn't exit non-zero, suppression is the correct fix
anyway.

Bug: None
Test: make lint
Change-Id: I92bb6ce6dcd13c41474b05e84c8518f72ca7611d