Handle case where upstream has no tags.
Now that we're checking for a tag that could alternatively be upgraded
to instead of a SHA, we might get into the situation where we're
looking for upstream tags in a repo that doesn't have any. `git
describe` fails in this case, and we're not handling that. Annoyingly
there's no specific return code for this case, so we need to match the
error text to detect this case.

In the case where the repo is tracking a SHA and there are no upstream
tags on the default branch, no suggestion to switch to tag will be
shown.

In the case where the repo is tracking a tag and there are no upstream
tags on the default branch, an error will be emitted, because
presumably that's a misconfigured METADATA. It's possible that we have
projects that are tracking tags from non-default branches that would
emit an error in this case, but I think those would have been broken
for other reasons (we're searching the wrong branch for tags), so that
should probably be cleaned up separately if needed.

Bug: None
Test: make test
Change-Id: I53fe1d5e32fce49b1a372e574ebd717c9d24659b