Fix bug in post_update of Archive projects
When an Archive project is upgraded, we unzip the new downloaded archive
and swap the current project with the old one. But we never return the
dir to where we save the old project. This causes post_update script to
be nonfunctional when it's looking for some pieces in the old dir.

Test: updater.sh update external/obstack
Change-Id: I6efb85245e9b5c455f9a5e727f8b56cf1e861460