Fix can not locate root of source tree error
Kokoro could successfully repo init and repo sync in the TOP dir it was given. But it could not locate the TOP from EXTERNAL_UPDATER when we cd'd into EXTERNAL_UPDATER. This is because we defined EXTERNAL_UPDATER relative to pwd but we are not sure where Kokoro thinks our writable working directory is and we made a wrong assumption. Since repo init and repo sync were successful in TOP, we are going to find external_updater's path relative to the TOP we defined.

Test: TreeHugger
Change-Id: Idcb60a900b74d5e1ea169ab40eb001181e6eae86