Improve validate parser
The validate parser only showed the diff of an external project with its
upstream. I added more useful information to display to the user.

Sometimes users manually set the wrong version in the METADATA file
which causes the validate parser to show a larger diff with upstream
than usual. We now check for the accuracy of the version in the METADATA
file and let the user know if the version is not accurate and what SHA
or tag it should be.

We sometimes see 'Git' in the METADATA file of an 'Archive'. We now also
let the user know that the project is an 'Archive'.

Bug: None
Test: TreeHugger
Change-Id: I710f1e4caaa4d0265da59caa196d462cf665dd05