Add support for updating from upstream's non-default branch
Instead of adding an extra field to METADATA proto or a new config file
for each project, we can change the url.value of Git projects to contain
the branch we want to pull data from. We can add support for git hosts
other than GitHub when we have requests from those repos.

Bug: 392999436
Test: TreeHugger
Change-Id: If90193d8c4e3b8a54029e244ab85004fc8e1bbbc